+++ app/Exercicio1/Adders/NumberFilter/BaseNumberFilter.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Onboarding\Adders\NumberFilter;
+namespace Onboarding\Exercicio1\Adders\NumberFilter;
 
 abstract class BaseNumberFilter
 {
