+++ app/Exercicio1/Adders/NumberFilter/ThreeAndFiveNumberFilter.php
@@ -1,18 +1,17 @@
 <?php
 
-namespace Onboarding\Adders\NumberFilter;
+namespace Onboarding\Exercicio1\Adders\NumberFilter;
 
-class ThreeOrFiveAndSevenNumberFilter extends BaseNumberFilter
+class ThreeAndFiveNumberFilter extends BaseNumberFilter
 {
     public function filterNumbers(array $arrayToFilter): array
     {
         return array_filter($arrayToFilter, function ($arrayNumber) {
             $numberIsMultipleOfThree = $this->numberIsMultiple(number: $arrayNumber, multiple: 3);
             $numberIsMultipleOfFive = $this->numberIsMultiple(number: $arrayNumber, multiple: 5);
-            $numberIsMultipleOfSeven = $this->numberIsMultiple(number: $arrayNumber, multiple: 7);
 
-            return ($numberIsMultipleOfThree || $numberIsMultipleOfFive)
-                && $numberIsMultipleOfSeven;
+            return $numberIsMultipleOfThree
+                && $numberIsMultipleOfFive;
         });
     }
 }
