+++ app/Exercicio1/Adders/NumberFilter/ThreeOrFiveAndSevenNumberFilter.php
@@ -1,17 +1,18 @@
 <?php
 
-namespace Onboarding\Adders\NumberFilter;
+namespace Onboarding\Exercicio1\Adders\NumberFilter;
 
-class ThreeAndFiveNumberFilter extends BaseNumberFilter
+class ThreeOrFiveAndSevenNumberFilter extends BaseNumberFilter
 {
     public function filterNumbers(array $arrayToFilter): array
     {
         return array_filter($arrayToFilter, function ($arrayNumber) {
             $numberIsMultipleOfThree = $this->numberIsMultiple(number: $arrayNumber, multiple: 3);
             $numberIsMultipleOfFive = $this->numberIsMultiple(number: $arrayNumber, multiple: 5);
+            $numberIsMultipleOfSeven = $this->numberIsMultiple(number: $arrayNumber, multiple: 7);
 
-            return $numberIsMultipleOfThree
-                && $numberIsMultipleOfFive;
+            return ($numberIsMultipleOfThree || $numberIsMultipleOfFive)
+                && $numberIsMultipleOfSeven;
         });
     }
 }
